requirements.txt
huggingface_hub
python-dotenv
pyarrow
jupyter
matplotlib
seaborn
nbformat
scikit-learn
streamlit
plotly
